Make bus error publication configurable
