+++ src/Abc.Zebus.Tests/Core/BusTests.Reply.cs
@@ -9,69 +9,72 @@ namespace Abc.Zebus.Tests.Core
 {
     public partial class BusTests
     {
-        [Test]
-        public void handlers_reply_with_an_int()
+        public class Replay : BusTests
         {
-            using (MessageId.PauseIdGeneration())
+            [Test]
+            public void handlers_reply_with_an_int()
             {
-                const int commandReply = 456;
-                var command = new FakeCommand(123);
-                SetupDispatch(command, _ => _bus.Reply(commandReply));
+                using (MessageId.PauseIdGeneration())
+                {
+                    const int commandReply = 456;
+                    var command = new FakeCommand(123);
+                    SetupDispatch(command, _ => _bus.Reply(commandReply));
 
-                var transportMessageReceived = command.ToTransportMessage(_peerUp);
-                var expectedTransportMessage = new MessageExecutionCompleted(transportMessageReceived.Id, commandReply, null).ToTransportMessage(_self);
+                    var transportMessageReceived = command.ToTransportMessage(_peerUp);
+                    var expectedTransportMessage = new MessageExecutionCompleted(transportMessageReceived.Id, commandReply, null).ToTransportMessage(_self);
 
-                _transport.RaiseMessageReceived(transportMessageReceived);
+                    _transport.RaiseMessageReceived(transportMessageReceived);
 
-                var sentMessage = _transport.Messages.Single();
-                expectedTransportMessage.ShouldHaveSamePropertiesAs(sentMessage.TransportMessage);
-                var destination = sentMessage.Targets.Single();
-                destination.ShouldHaveSamePropertiesAs(_peerUp);
+                    var sentMessage = _transport.Messages.Single();
+                    expectedTransportMessage.ShouldHaveSamePropertiesAs(sentMessage.TransportMessage);
+                    var destination = sentMessage.Targets.Single();
+                    destination.ShouldHaveSamePropertiesAs(_peerUp);
+                }
             }
-        }
 
-        [Test]
-        public void handlers_reply_with_an_int_and_a_message()
-        {
-            using (MessageId.PauseIdGeneration())
+            [Test]
+            public void handlers_reply_with_an_int_and_a_message()
             {
-                const int commandReply = 456;
-                const string replyMessage = "Test";
+                using (MessageId.PauseIdGeneration())
+                {
+                    const int commandReply = 456;
+                    const string replyMessage = "Test";
 
-                var command = new FakeCommand(123);
-                SetupDispatch(command, _ => _bus.Reply(commandReply, replyMessage));
+                    var command = new FakeCommand(123);
+                    SetupDispatch(command, _ => _bus.Reply(commandReply, replyMessage));
 
-                var transportMessageReceived = command.ToTransportMessage(_peerUp);
-                var expectedTransportMessage = new MessageExecutionCompleted(transportMessageReceived.Id, commandReply, replyMessage).ToTransportMessage(_self);
+                    var transportMessageReceived = command.ToTransportMessage(_peerUp);
+                    var expectedTransportMessage = new MessageExecutionCompleted(transportMessageReceived.Id, commandReply, replyMessage).ToTransportMessage(_self);
 
-                _transport.RaiseMessageReceived(transportMessageReceived);
+                    _transport.RaiseMessageReceived(transportMessageReceived);
 
-                var sentMessage = _transport.Messages.Single();
-                expectedTransportMessage.ShouldHaveSamePropertiesAs(sentMessage.TransportMessage);
-                var destination = sentMessage.Targets.Single();
-                destination.ShouldHaveSamePropertiesAs(_peerUp);
+                    var sentMessage = _transport.Messages.Single();
+                    expectedTransportMessage.ShouldHaveSamePropertiesAs(sentMessage.TransportMessage);
+                    var destination = sentMessage.Targets.Single();
+                    destination.ShouldHaveSamePropertiesAs(_peerUp);
+                }
             }
-        }
 
-        [Test]
-        public void handlers_reply_with_an_object()
-        {
-            using (MessageId.PauseIdGeneration())
+            [Test]
+            public void handlers_reply_with_an_object()
             {
-                var command = new FakeCommand(123);
-                var commandResult = new FakeCommandResult("CommandResult", 45);
-                var transportMessageReceived = command.ToTransportMessage(_peerUp);
-                SetupDispatch(command, _ => _bus.Reply(commandResult));
+                using (MessageId.PauseIdGeneration())
+                {
+                    var command = new FakeCommand(123);
+                    var commandResult = new FakeCommandResult("CommandResult", 45);
+                    var transportMessageReceived = command.ToTransportMessage(_peerUp);
+                    SetupDispatch(command, _ => _bus.Reply(commandResult));
 
-                var expectedExecutionCompleted = MessageExecutionCompleted.Success(transportMessageReceived.Id, commandResult, _messageSerializer);
-                var expectedTransportMessage = expectedExecutionCompleted.ToTransportMessage(_self);
+                    var expectedExecutionCompleted = MessageExecutionCompleted.Success(transportMessageReceived.Id, commandResult, _messageSerializer);
+                    var expectedTransportMessage = expectedExecutionCompleted.ToTransportMessage(_self);
 
-                _transport.RaiseMessageReceived(transportMessageReceived);
+                    _transport.RaiseMessageReceived(transportMessageReceived);
 
-                var sentMessage = _transport.Messages.Single();
-                expectedTransportMessage.ShouldHaveSamePropertiesAs(sentMessage.TransportMessage);
-                var destination = sentMessage.Targets.Single();
-                destination.ShouldHaveSamePropertiesAs(_peerUp);
+                    var sentMessage = _transport.Messages.Single();
+                    expectedTransportMessage.ShouldHaveSamePropertiesAs(sentMessage.TransportMessage);
+                    var destination = sentMessage.Targets.Single();
+                    destination.ShouldHaveSamePropertiesAs(_peerUp);
+                }
             }
         }
     }
